+++ src/vect3.c
@@ -20,21 +20,21 @@ Vect3 Vect3Neg(Vect3 v) {
     return v;
 }
 
-void Vect3ScaleP(Vect3* v1, double s) {
+void Vect3ScaleP(Vect3* v1, float s) {
     v1->x *= s;
     v1->y *= s;
     v1->z *= s;
 }
 
-Vect3 Vect3Scale(Vect3 v1, double s) {
+Vect3 Vect3Scale(Vect3 v1, float s) {
     v1.x *= s;
     v1.y *= s;
     v1.z *= s;
     return v1;
 }
 
-double Vect3Len(Vect3 v1) {
-    double sum = 0;
+float Vect3Len(Vect3 v1) {
+    float sum = 0;
     sum += v1.x;
     sum += v1.y;
     sum += v1.z;
@@ -44,7 +44,7 @@ double Vect3Len(Vect3 v1) {
 }
 
 void Vect3NormP(Vect3* v1) {
-    double sum = sqrt(sum);
+    float sum = sqrt(sum);
 
     v1->x /= sum;
     v1->y /= sum;
@@ -52,7 +52,7 @@ void Vect3NormP(Vect3* v1) {
 }
 
 Vect3 Vect3Norm(Vect3 v1) {
-    double sum = Vect3Len(v1);
+    float sum = Vect3Len(v1);
 
     v1.x /= sum;
     v1.y /= sum;
@@ -60,7 +60,7 @@ Vect3 Vect3Norm(Vect3 v1) {
     return v1;
 }
 
-double Vect3Dot(Vect3 v1, Vect3 v2) {
+float Vect3Dot(Vect3 v1, Vect3 v2) {
     return (v1.x * v2.x) + 
            (v1.y * v2.y) + 
            (v1.z * v2.z);
